change namespace from Spid::Rails to SpidRails
