+++ app/assets/javascripts/spid_rails/application.js
