+++ app/assets/javascripts/spid_rails/metadata.js
