+++ app/assets/stylesheets/spid_rails/application.css
