+++ app/assets/stylesheets/spid_rails/metadata.css
