--- app/controllers/spid/rails/metadata_controller.rb
@@ -1,10 +1,0 @@
-require_dependency "spid/rails/application_controller"
-
-# Metadata for Service Provider
-class Spid::Rails::MetadataController < Spid::Rails::SpidController
-  def show
-    metadata = Spid::Rails::Metadata.create(host: main_app.root_url)
-    render xml: metadata.to_xml
-  end
-
-end
