+++ app/controllers/spid_rails/application_controller.rb
@@ -1,0 +1,5 @@
+module SpidRails
+  class ApplicationController < ActionController::Base
+    protect_from_forgery with: :exception
+  end
+end
