+++ app/controllers/spid_rails/metadata_controller.rb
@@ -1,0 +1,10 @@
+require_dependency "spid_rails/application_controller"
+
+# Metadata del Service Provider
+class SpidRails::MetadataController < SpidRails::SpidController
+  def show
+    metadata = SpidRails::Metadata.create(host: main_app.root_url)
+    render xml: metadata.to_xml
+  end
+
+end
