--- app/helpers/spid/rails/application_helper.rb
@@ -1,6 +1,0 @@
-module Spid
-  module Rails
-    module ApplicationHelper
-    end
-  end
-end
