--- app/helpers/spid/rails/metadata_helper.rb
@@ -1,4 +1,0 @@
-module Spid::Rails
-  module MetadataHelper
-  end
-end
