+++ app/helpers/spid_rails/application_helper.rb
@@ -1,0 +1,4 @@
+module SpidRails
+  module ApplicationHelper
+  end
+end
