--- app/jobs/spid/rails/application_job.rb
@@ -1,6 +1,0 @@
-module Spid
-  module Rails
-    class ApplicationJob < ActiveJob::Base
-    end
-  end
-end
