+++ app/jobs/spid_rails/application_job.rb
@@ -1,0 +1,4 @@
+module SpidRails
+  class ApplicationJob < ActiveJob::Base
+  end
+end
