--- app/mailers/spid/rails/application_mailer.rb
@@ -1,8 +1,0 @@
-module Spid
-  module Rails
-    class ApplicationMailer < ActionMailer::Base
-      default from: 'from@example.com'
-      layout 'mailer'
-    end
-  end
-end
