+++ app/mailers/spid_rails/application_mailer.rb
@@ -1,0 +1,6 @@
+module SpidRails
+  class ApplicationMailer < ActionMailer::Base
+    default from: 'from@example.com'
+    layout 'mailer'
+  end
+end
