--- app/models/spid/rails/application_record.rb
@@ -1,7 +1,0 @@
-module Spid
-  module Rails
-    class ApplicationRecord < ActiveRecord::Base
-      self.abstract_class = true
-    end
-  end
-end
