--- app/models/spid/rails/settings/metadata.rb
@@ -1,13 +1,0 @@
-module Spid
-  module Rails
-
-    class Settings::Metadata < Settings
-
-      def to_hash
-        sp_attributes
-      end
-
-    end
-
-  end
-end
