+++ app/models/spid_rails/application_record.rb
@@ -1,0 +1,5 @@
+module SpidRails
+  class ApplicationRecord < ActiveRecord::Base
+    self.abstract_class = true
+  end
+end
