+++ app/models/spid_rails/settings/metadata.rb
@@ -1,0 +1,11 @@
+module SpidRails
+
+  class Settings::Metadata < Settings
+
+    def to_hash
+      sp_attributes
+    end
+
+  end
+  
+end
