--- app/views/layouts/spid/rails/application.html.erb
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Spid rails</title>
-  <%= stylesheet_link_tag    "spid/rails/application", media: "all" %>
-  <%= javascript_include_tag "spid/rails/application" %>
-  <%= csrf_meta_tags %>
-</head>
-<body>
-
-<%= yield %>
-
-</body>
-</html>
