+++ app/views/layouts/spid_rails/application.html.erb
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Spid rails</title>
+  <%= stylesheet_link_tag    "spid_rails/application", media: "all" %>
+  <%= javascript_include_tag "spid_rails/application" %>
+  <%= csrf_meta_tags %>
+</head>
+<body>
+
+<%= yield %>
+
+</body>
+</html>
