+++ lib/generators/spid_rails/templates/spid_rails.rb
@@ -1,6 +1,6 @@
 # Impostazioni di default dello Spid Engine
 
-Spid::Rails.tap do |config|
+SpidRails.tap do |config|
 
   # Mount point di Spid sull'applicazione
   # default: 'spid'
