--- lib/spid/rails/engine.rb
@@ -1,10 +1,0 @@
-require 'haml-rails'
-require 'onelogin/ruby-saml'
-
-module Spid
-  module Rails
-    class Engine < ::Rails::Engine
-      isolate_namespace Spid::Rails
-    end
-  end
-end
