--- lib/spid/rails/version.rb
@@ -1,5 +1,0 @@
-module Spid
-  module Rails
-    VERSION = '0.1.0'
-  end
-end
