+++ lib/spid_rails/engine.rb
@@ -1,0 +1,8 @@
+require 'haml-rails'
+require 'onelogin/ruby-saml'
+
+module SpidRails
+  class Engine < ::Rails::Engine
+    isolate_namespace SpidRails
+  end
+end
