+++ lib/spid_rails/version.rb
@@ -1,0 +1,3 @@
+module SpidRails
+  VERSION = '0.1.0'
+end
