+++ test/spid_rails_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class SpidRails::Test < ActiveSupport::TestCase
+  test "truth" do
+    assert_kind_of Module, SpidRails
+  end
+end
